+++ generators/gen_client_cli_md.py
@@ -2,13 +2,13 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
+from pathlib import Path
 from typing import Final
 
 import typer.core
-from pathlib import Path
 from click.testing import CliRunner
-from s2gos.client.cli.cli import cli
 
+from s2gos.client.cli.cli import cli
 
 DOCS_PATH: Final = (Path(__file__).parent / ".." / "docs").resolve()
 OUTPUT_FILE: Final = DOCS_PATH / "cli.md"