+++ s2gos/server/main.py
@@ -2,13 +2,27 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-from .app import app
-from . import routes
+import importlib
+import os
 
+from . import routes
+from .app import app
 from .impl.local_service import LocalService
 from .provider import ServiceProvider
 
-# TODO: should use env var (in .env) to configure actual service to be used
-ServiceProvider.set_instance(LocalService())
+
+def get_service():
+    service_class_spec = os.environ.get("S2GOS_SERVICE_CLASS")
+    if service_class_spec is None:
+        # TODO: replace by true S2GOS service
+        return LocalService()
+    else:
+        module_name, class_name = service_class_spec.split(":", maxsplit=1)
+        module = importlib.import_module(module_name)
+        service_cls = getattr(module, class_name)
+        return service_cls()
+
+
+ServiceProvider.set_instance(get_service())
 
 __all__ = ["app", "routes"]