+++ CHANGES.md
@@ -1,8 +1,13 @@
-## Changes in version 0.0.2 (in development)
+## Changes in version 0.0.3 (in development)
+
+## Changes in version 0.0.2 (not released)
 
 * Reorganized package
 * Started code generation for models, client, and server
 * Added basic FastAPI server, added server CLI `s2gos-server`
+* Using dummy implementation for the S2GOS service
+* Using `Typer` instead of `Click`
+* Added pretty rendering of model object in Jupyter notebooks
 
 ## Version 0.0.1 (not released)